Refactor back to top functionality
Instead of placing the backToTop fab icon in a fixed position and having enough bottom margin to not be overlapped with the Footer, I decided to make it sticky and place it at the bottom of the HomeContainer. A dirty negative top Margin is needed to pull down the last section.